# perf(execute): optimize group lookup for non ordered access patterns