--- a/execute/group_lookup_test.go
+++ b/execute/group_lookup_test.go
@@ -1,13 +1,16 @@
 package execute_test
 
 import (
+	"math/rand"
 	"reflect"
 	"sort"
 	"testing"
+	"time"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/influxdata/flux"
 	"github.com/influxdata/flux/execute"
+	"github.com/influxdata/flux/internal/gen"
 	"github.com/influxdata/flux/semantic"
 	"github.com/influxdata/flux/values"
 )
@@ -216,4 +219,147 @@ func TestGroupLookup_RangeWithDelete(t *testing.T) {
 type entry struct {
 	Key   flux.GroupKey
 	Value int
+}
+
+func testGroupLookup_LookupOrCreate(tb testing.TB, run func(name string, keys []flux.GroupKey)) {
+	// Generate a small schema to use for the group lookup.
+	schema := gen.Schema{
+		Start: time.Now(),
+		Tags: []gen.Tag{ // total cardinality is 105
+			{Name: "_measurement", Cardinality: 1},
+			{Name: "_field", Cardinality: 5},
+			{Name: "t0", Cardinality: 3},
+			{Name: "t1", Cardinality: 7},
+		},
+		NumPoints: 1,
+		Types: map[flux.ColType]int{
+			flux.TFloat: 1,
+		},
+		Seed: func(x int64) *int64 { return &x }(0),
+	}
+	input, err := gen.Input(schema)
+	if err != nil {
+		tb.Fatalf("unexpected error: %s", err)
+	}
+
+	// There should be one point per key so read from the input
+	// and put them into a slice.
+	canonicalKeys := make([]flux.GroupKey, 0, 105)
+	for input.More() {
+		res := input.Next()
+		if err := res.Tables().Do(func(table flux.Table) error {
+			canonicalKeys = append(canonicalKeys, table.Key())
+			return nil
+		}); err != nil {
+			tb.Fatalf("unexpected error: %s", err)
+		}
+	}
+
+	if len(canonicalKeys) != 105 {
+		tb.Fatalf("unexpected number of keys: %d", len(canonicalKeys))
+	}
+
+	run("Ordered", func() []flux.GroupKey {
+		keys := make([]flux.GroupKey, len(canonicalKeys))
+		copy(keys, canonicalKeys)
+		sort.Sort(flux.GroupKeys(keys))
+		return keys
+	}())
+
+	run("Interweaved", func() []flux.GroupKey {
+		keys := make([]flux.GroupKey, len(canonicalKeys))
+		copy(keys, canonicalKeys)
+		sort.Sort(flux.GroupKeys(keys))
+
+		// Fisher-yates shuffle with groups of 5.
+		tmp := make([]flux.GroupKey, 5)
+		for i := 100; i > 0; i -= 5 {
+			j := rand.Intn(i/5+1) * 5
+			if i == j {
+				continue
+			}
+			// Copy the current group into the tmp slice.
+			copy(tmp, keys[i:i+5])
+			// Copy the chosen range into the current group.
+			copy(keys[i:], keys[j:j+5])
+			// Copy the temporary keys into the chosen one.
+			copy(keys[j:], tmp)
+		}
+		return keys
+	}())
+
+	run("Unordered", func() []flux.GroupKey {
+		keys := make([]flux.GroupKey, len(canonicalKeys))
+		copy(keys, canonicalKeys)
+
+		// Fisher-yates shuffle.
+		for i := len(keys) - 1; i > 0; i-- {
+			j := rand.Intn(i + 1)
+			keys[i], keys[j] = keys[j], keys[i]
+		}
+		return keys
+	}())
+
+	run("Reversed", func() []flux.GroupKey {
+		keys := make([]flux.GroupKey, len(canonicalKeys))
+		copy(keys, canonicalKeys)
+
+		// Reverse the array.
+		for i, j := 0, len(keys)-1; i < j; i, j = i+1, j-1 {
+			keys[i], keys[j] = keys[j], keys[i]
+		}
+		return keys
+	}())
+
+	run("ManySplits", func() []flux.GroupKey {
+		keys := make([]flux.GroupKey, len(canonicalKeys))
+		copy(keys, canonicalKeys)
+
+		// Swap every 2nd element with every 3rd element to
+		// force a split to occur.
+		for i := 0; i < len(keys)-2; i += 3 {
+			keys[i+1], keys[i+2] = keys[i+2], keys[i+1]
+		}
+		return keys
+	}())
+}
+
+func TestGroupLookup_LookupOrCreate(t *testing.T) {
+	testGroupLookup_LookupOrCreate(t, func(name string, keys []flux.GroupKey) {
+		t.Run(name, func(t *testing.T) {
+			l := execute.NewGroupLookup()
+			for _, key := range keys {
+				if _, ok := l.Lookup(key); !ok {
+					l.Set(key, true)
+				} else {
+					t.Errorf("unexpected key lookup: %s", key)
+				}
+			}
+
+			// Lookup should work on everything.
+			for _, key := range keys {
+				if _, ok := l.Lookup(key); !ok {
+					t.Errorf("key lookup failed: %s", key)
+				}
+			}
+		})
+	})
+}
+
+func BenchmarkGroupLookup_LookupOrCreate(b *testing.B) {
+	testGroupLookup_LookupOrCreate(b, func(name string, keys []flux.GroupKey) {
+		b.Run(name, func(b *testing.B) {
+			b.ResetTimer()
+			b.ReportAllocs()
+
+			for i := 0; i < b.N; i++ {
+				l := execute.NewGroupLookup()
+				for _, key := range keys {
+					if _, ok := l.Lookup(key); !ok {
+						l.Set(key, true)
+					}
+				}
+			}
+		})
+	})
 }