This optimizes group lookup for the common scenario of ordered and
interweaved additions with some benefits to unordered. It reduces the
number of allocations and the usage of `sort.Search` by creating groups of
entries that are appended to.

The data structure contains a slice of time ranges. When adding a new
value, it will find which of the time ranges it can append to easily. So
when we are adding new values in sorted order, this results in a single
slice that is continuously appended to.

If the new value is sorted before all of the existing groups, it will
create a new group at the beginning and add the value to that group instead
of attempting to shift all of the values in the existing group like you
would have to do for a single slice.

If the new value would get inserted into the middle of a group such as with
unordered access, it will split the existing group into two so the values
after the currently added value would be in a new group and the new value
can be appended to the current group.

This means that if you add values in a sorted order, you will only get a
worst case scenario the first time you append a value rather than on every
insertion.

The worst case scenario for this data structure is adding the values in a
reverse sorted order.

The data structure could further be improved by having the top slice be
sorted backwards rather than sorted in the current direction. This would
likely eliminate the worst case scenario and would prevent the copy when
appending a key that is less than all of the other keys.

- [x] docs/SPEC.md updated
- [x] Test cases written

Fixes #1035.