@jsternberg 2019-03-12
Here are the performance numbers:

```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkGroupLookup_LookupOrCreate/Ordered-12         129291        52930         -59.06%
BenchmarkGroupLookup_LookupOrCreate/Interweaved-12     3113840       85321         -97.26%
BenchmarkGroupLookup_LookupOrCreate/Unordered-12       2863064       193151        -93.25%

benchmark                                              old allocs     new allocs     delta
BenchmarkGroupLookup_LookupOrCreate/Ordered-12         2              10             +400.00%
BenchmarkGroupLookup_LookupOrCreate/Interweaved-12     86             35             -59.30%
BenchmarkGroupLookup_LookupOrCreate/Unordered-12       102            152            +49.02%

benchmark                                              old bytes     new bytes     delta
BenchmarkGroupLookup_LookupOrCreate/Ordered-12         9728          10648         +9.46%
BenchmarkGroupLookup_LookupOrCreate/Interweaved-12     12416         9656          -22.23%
BenchmarkGroupLookup_LookupOrCreate/Unordered-12       12928         11784         -8.85%
```

As can be seen, there's some additional overhead for the new data structure in the sorted method, but the CPU usage is universally better across the board. The additional allocations are likely because I am not preallocating array space for the added entries and they don't appear to be too significant.

We will likely have to decide if the memory allocations are worth it. The number of bytes appears to be less for everything except ordered where ordered is likely because the test uses 105 series and the old group lookup had an initial buffer of 100 so it only needed to do one additional allocation.