@jsternberg 2019-03-13
I'll integrate these into the pushed PR, but here's two additional benchmarks. One is if the groups are added in reverse order and one is where I believe there will be many splits. It takes sorted groups of 3 and swaps the 2nd and 3rd elements. This means that the 1st gets appended somewhere, the 3rd gets appended, and then the 2nd forces a split.

```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkGroupLookup_LookupOrCreate/Ordered-12         130501        52601         -59.69%
BenchmarkGroupLookup_LookupOrCreate/Interweaved-12     2954610       83837         -97.16%
BenchmarkGroupLookup_LookupOrCreate/Unordered-12       2932631       176408        -93.98%
BenchmarkGroupLookup_LookupOrCreate/Reversed-12        3947501       186467        -95.28%
BenchmarkGroupLookup_LookupOrCreate/ManySplits-12      1016825       129218        -87.29%

benchmark                                              old allocs     new allocs     delta
BenchmarkGroupLookup_LookupOrCreate/Ordered-12         2              10             +400.00%
BenchmarkGroupLookup_LookupOrCreate/Interweaved-12     86             35             -59.30%
BenchmarkGroupLookup_LookupOrCreate/Unordered-12       102            152            +49.02%
BenchmarkGroupLookup_LookupOrCreate/Reversed-12        105            218            +107.62%
BenchmarkGroupLookup_LookupOrCreate/ManySplits-12      36             148            +311.11%

benchmark                                              old bytes     new bytes     delta
BenchmarkGroupLookup_LookupOrCreate/Ordered-12         9728          10648         +9.46%
BenchmarkGroupLookup_LookupOrCreate/Interweaved-12     12416         9656          -22.23%
BenchmarkGroupLookup_LookupOrCreate/Unordered-12       12928         11784         -8.85%
BenchmarkGroupLookup_LookupOrCreate/Reversed-12        13024         12120         -6.94%
BenchmarkGroupLookup_LookupOrCreate/ManySplits-12      10816         12712         +17.53%
```

Memory consumption is a bit higher, but CPU usage is still down. I think it might help to preallocate a small array for each new key list, but I'll save that for another PR.