@jsternberg 2019-03-13
I refactored the tests and benchmarks so they would share code. If it passes CI, I'll merge this and we can iterate on it further.